fix time

+++ Services/BaseService.cs
@@ -42,7 +42,7 @@ namespace Coflnet.Sky.BFCS.Services
                     return;
                 prod.Publish("snipes", MessagePack.MessagePackSerializer.Serialize(lp), CommandFlags.FireAndForget);
                 Console.WriteLine($"found {lp.Finder} :O {lp.Auction.Uuid} {lp.Auction.ItemName}");
-                var timestamp = DateTime.Now.Subtract(new DateTime(1970, 1, 1)).TotalMilliseconds;
+                var timestamp = (long)DateTime.Now.Subtract(new DateTime(1970, 1, 1)).TotalMilliseconds;
                 Task.Run(async ()=>{
                     await httpClient.PostAsync($"https://sky.coflnet.com/api/flip/track/found/{lp.Auction.Uuid}?finder=test&price={lp.TargetPrice}&timeStamp={timestamp}", null);
                 });
